On this project I will work on these Analysis:  How long usually a video can trend in different countries? which category has longer trend? Most common words(Trending words) in video titles which category has highest ratio of Likes-Dislikes and Views-Comments.
This is the project I wanted to focus on: sentiment analysis and how YouTubers can actually make the most use of their data, like the comments they receive on each video they publish. It took me a month to wrap up this project, and I believe the outcome is brilliant as it showcases all the details of sentimental behaviors behind the scenes. This highlights the power of data science.